+++ src/pysatl_criterion/hypothesis_testing/critical_values/cv_calculator/cv_calculator.py
@@ -1,11 +1,11 @@
 import numpy as np
 import scipy.stats as scipy_stats
 
-from src.pysatl_criterion.persistence.models.limit_distribution import (
+from pysatl_criterion.persistence.models.limit_distribution import (
     CriticalValueQuery,
     ILimitDistributionStorage,
 )
-from src.pysatl_criterion.statistics.models import HypothesisType
+from pysatl_criterion.statistics.models import HypothesisType
 
 
 class CVCalculator:
