+++ src/pysatl_criterion/hypothesis_testing/critical_values/loader/remote_loader.py
@@ -1,6 +1,6 @@
 import logging
 
-from src.pysatl_criterion.persistence.models.limit_distribution import (
+from pysatl_criterion.persistence.models.limit_distribution import (
     BulkLoadResult,
     CriticalValueQuery,
     ILimitDistributionStorage,
