+++ src/pysatl_criterion/hypothesis_testing/critical_values/resolver/model.py
@@ -1,7 +1,7 @@
 from abc import ABC, abstractmethod
 
-from src.pysatl_criterion.hypothesis_testing.critical_values.critical_area.model import CriticalArea
-from src.pysatl_criterion.statistics.models import HypothesisType
+from pysatl_criterion.hypothesis_testing.critical_values.critical_area.model import CriticalArea
+from pysatl_criterion.statistics.models import HypothesisType
 
 
 class CriticalValueResolver(ABC):
