+++ src/pysatl_criterion/hypothesis_testing/p_value/resolver/calculation_resolver.py
@@ -1,11 +1,11 @@
 import scipy.stats as scipy_stats
 from typing_extensions import override
 
-from src.pysatl_criterion.persistence.models.limit_distribution import (
+from pysatl_criterion.persistence.models.limit_distribution import (
     CriticalValueQuery,
     ILimitDistributionStorage,
 )
-from src.pysatl_criterion.statistics.models import HypothesisType
+from pysatl_criterion.statistics.models import HypothesisType
 
 from .model import PValueResolver
