+++ src/pysatl_criterion/hypothesis_testing/p_value/resolver/model.py
@@ -1,6 +1,6 @@
 from abc import ABC, abstractmethod
 
-from src.pysatl_criterion.statistics.models import HypothesisType
+from pysatl_criterion.statistics.models import HypothesisType
 
 
 class PValueResolver(ABC):
