+++ src/pysatl_criterion/persistence/sqlalchemy/models/limit_distribution_orm.py
@@ -3,8 +3,8 @@ import json
 from sqlalchemy import PrimaryKeyConstraint, Text
 from sqlalchemy.orm import Mapped, declarative_base, mapped_column
 
-from src.pysatl_criterion.persistence.models.limit_distribution import LimitDistributionModel
-from src.pysatl_criterion.persistence.sqlalchemy.alchemy_decorator import CompressedFloatArray
+from pysatl_criterion.persistence.models.limit_distribution import LimitDistributionModel
+from pysatl_criterion.persistence.sqlalchemy.alchemy_decorator import CompressedFloatArray
 
 
 Base = declarative_base()
