+++ src/pysatl_criterion/statistics/normal.py
@@ -6,9 +6,9 @@ import pandas as pd
 import scipy.stats as scipy_stats
 from typing_extensions import override
 
-from src.pysatl_criterion.statistics.common import ADStatistic, KSStatistic, LillieforsTest
-from src.pysatl_criterion.statistics.goodness_of_fit import AbstractGoodnessOfFitStatistic
-from src.pysatl_criterion.statistics.graph_goodness_of_fit import (
+from pysatl_criterion.statistics.common import ADStatistic, KSStatistic, LillieforsTest
+from pysatl_criterion.statistics.goodness_of_fit import AbstractGoodnessOfFitStatistic
+from pysatl_criterion.statistics.graph_goodness_of_fit import (
     AbstractGraphTestStatistic,
     GraphAverageDegreeTestStatistic,
     GraphCliqueNumberTestStatistic,
