+++ src/pysatl_criterion/statistics/student.py
@@ -4,8 +4,12 @@ import numpy as np
 import scipy.stats as scipy_stats
 from typing_extensions import override
 
-from src.pysatl_criterion.statistics.common import ADStatistic, CrammerVonMisesStatistic, KSStatistic
-from src.pysatl_criterion.statistics.goodness_of_fit import AbstractGoodnessOfFitStatistic
+from pysatl_criterion.statistics.common import (
+    ADStatistic,
+    CrammerVonMisesStatistic,
+    KSStatistic,
+)
+from pysatl_criterion.statistics.goodness_of_fit import AbstractGoodnessOfFitStatistic
 
 
 class AbstractStudentGofStatistic(AbstractGoodnessOfFitStatistic, ABC):
