+++ tests/distributions/test_beta.py
@@ -2,7 +2,7 @@ import numpy as np
 import pytest
 import scipy.stats as scipy_stats
 
-from src.pysatl_criterion.statistics.beta import (
+from pysatl_criterion.statistics.beta import (
     AbstractBetaGofStatistic,
     AndersonDarlingBetaGofStatistic,
     Chi2PearsonBetaGofStatistic,
