+++ tests/distributions/test_student.py
@@ -9,7 +9,7 @@ import numpy as np
 import pytest
 import scipy.stats as scipy_stats
 
-from src.pysatl_criterion.statistics.student import (
+from pysatl_criterion.statistics.student import (
     AbstractStudentGofStatistic,
     AndersonDarlingStudentGofStatistic,
     ChiSquareStudentGofStatistic,
