+++ tests/distributions/test_weibull.py
@@ -1,6 +1,6 @@
 import pytest
 
-from src.pysatl_criterion.statistics.weibull import (
+from pysatl_criterion.statistics.weibull import (
     AndersonDarlingWeibullGofStatistic,
     CrammerVonMisesWeibullGofStatistic,
     KolmogorovSmirnovWeibullGofStatistic,
