+++ tests/hypothesis_testing/critical_values/test_cv.py
@@ -3,10 +3,10 @@ from unittest.mock import MagicMock
 import numpy as np
 import pytest
 
-from src.pysatl_criterion.hypothesis_testing.critical_values.cv_calculator.cv_calculator import (
+from pysatl_criterion.hypothesis_testing.critical_values.cv_calculator.cv_calculator import (
     CVCalculator,
 )
-from src.pysatl_criterion.statistics.models import HypothesisType
+from pysatl_criterion.statistics.models import HypothesisType
 
 
 def test_calc_critical_value_right_tailed():
