+++ tests/hypothesis_testing/critical_values/test_remote_loader.py
@@ -1,14 +1,14 @@
 import numpy as np
 import pytest
 
-from src.pysatl_criterion.hypothesis_testing.critical_values.loader.remote_loader import (
+from pysatl_criterion.hypothesis_testing.critical_values.loader.remote_loader import (
     CriticalValueLoader,
 )
-from src.pysatl_criterion.persistence.models.limit_distribution import (
+from pysatl_criterion.persistence.models.limit_distribution import (
     CriticalValueQuery,
     LimitDistributionModel,
 )
-from src.pysatl_criterion.persistence.sqlalchemy.datastorage import AlchemyLimitDistributionStorage
+from pysatl_criterion.persistence.sqlalchemy.datastorage import AlchemyLimitDistributionStorage
 
 
 @pytest.fixture
