+++ tests/hypothesis_testing/goodness_of_fit_test/test_graph_goodness_of_fit.py
@@ -1,6 +1,6 @@
 import pytest
 
-from src.pysatl_criterion.statistics.graph_goodness_of_fit import AbstractGraphTestStatistic
+from pysatl_criterion.statistics.graph_goodness_of_fit import AbstractGraphTestStatistic
 
 
 def test_abstract_graph_criterion():
