+++ tests/hypothesis_testing/multiple_testing/test_fwer.py
@@ -1,6 +1,6 @@
 import pytest as pytest
 
-from src.pysatl_criterion import BonferroniMultipleTesting, Holm, SidakHolm, SidakMultipleTesting
+from pysatl_criterion import BonferroniMultipleTesting, Holm, SidakHolm, SidakMultipleTesting
 
 
 METHODS = [BonferroniMultipleTesting, SidakMultipleTesting]
