+++ tests/hypothesis_testing/p_value/resolver/test_calculation_resolver.py
@@ -10,10 +10,10 @@ from unittest.mock import MagicMock, patch
 import numpy as np
 import pytest
 
-from src.pysatl_criterion.hypothesis_testing.p_value.resolver.calculation_resolver import (
+from pysatl_criterion.hypothesis_testing.p_value.resolver.calculation_resolver import (
     CalculationPValueResolver,
 )
-from src.pysatl_criterion.statistics.models import HypothesisType
+from pysatl_criterion.statistics.models import HypothesisType
 
 
 @patch(
